Comment out RZGetCOMDllDirector definition
This is supposed to be user-defined. Remove our own
definition to make it easier to clone and extend this
repo.

All sample directors should eventually be moved to their
own repository. Obviously removing this definition will
break linkage when trying to compile this repository
standalone, which isn't desirable but it's fine if this
repo is extended and another library provides it anyway.