Fix the cIGZSystemService definitions
OnTick and OnIdle take a single integer parameter, which appears to be
some kind of incrementing counter. Not sure what the value is used for,
providing callers an easy way to track the time elapsed between calls?

The header was also missing the GetServiceTickPriority method, which
was causing it to crash when calling AddToTick.