Convert the cISC4* types over to the EASTL vector
The EASTL vector is binary compatible with the internal vector type
used by SC4.

Our EASTL vector uses a custom allocator that wraps SC4's internal
memory allocator.
This allows the vector to interoperate with SC4, but prevents it
from being used when the cIGZAllocatorService is not available.
For example, the EASTL vector could not be used after SC4 shuts down
the cIGZAllocatorService.

Because of this limitation the cRZCOMDllDirector class still uses
the vector implementation from MSVC.

This commit also fixes a build break with the cRZBaseVector class.
It requires a language version of at least C++ 17, so the project now
targets C++ 20.