Add SC4Preferences and related classes
The game stores its preferences in a class that contains nested classes
for the various preference categories.
The whole thing adds up to a 5940 byte SC4Preferences class that
behaves like a C struct.

This design was presumably chosen for performance reasons, as it allows
the game to read and write the preferences class as a binary blob.