Add a base class for cIGZUnknown types that delete themselves
This class handles the reference counting and deletes the object when
the reference count drops to zero.
It is intended to be used by GZCOM types that have instances created
on demand, e.g. any class that a DLL director registers using
cRZCOMDllDirector::AddCls(uint32_t,FactoryFunctionPtr1).

The cIGZUnknown-derived types that do not delete themselves when the
reference count drops to zero (e.g. DLL Directors) will have to
implement the reference counting on their own.