Update cISC4CivicBuildingSimulator::ConditionalBuildingStatus
Added the status flag values and filled in the unknown fields.
The unknown fields were a the item tool tip using the cRZString type
that provides SC4's internal implementation of the cIGZString
interface.

To support structures that contain cRZString fields and methods
that return it, a SC4String class has been added.
The SC4String class recreates the internal layout of cRZString
and provides methods to cast it to cIGZString.
The SC4String class will not work for the methods that expect a
cRZString to be passes as a reference parameter.