Add 2 missing functions the cISC4App
This allows DLLs to use the Get*Directory functions.

The first missing function is a method that returns the path of the
Exception Reports directory as a cRZString.
The second method returns a pointer or reference to an unknown type,
possibly an array. In the debugger memory view the data doesn't look
like a C++ class vtable.