Add a SC4CellRegion implementation
This class is used by cISC4ZoneManager::PlaceZone and other APIs that
require information on which cells they should modify.

SC4CellRegion acts as a Boolean mask, each cell in the bounds rectangle
has a corresponding bit in the mask data that tells the game if that
cell should be modified.
Any cell that has a mask bit value of 0 (false) will be left untouched
by the requested operation.