Update SC4Vector for arrays of cIGZUnknown
When the type is a class that is derived from cIGZUnknown, it needs to
be freed by calling the Release method.
The game's code will have called AddRef after adding the pointer to the
output vector.

This is necessary to correctly handle the cISC4DepartmentBudget* vector
that cISC4BudgetSimulator::GetDepartmentBudgetsInGroup fills.