Add a base class for new view input controls
This class uses the same default function return values as SC4's
internal implementation, and re-implements most of the helper
methods.

Derived classes should override the necessary On* methods to handle
input processing for their control.
If Init and Shutdown are overridden, the base class implementation
must be called first.
IsSelfScrollingView, ShouldStack, Activate, and/or Deactivate can also
be overridden if necessary.