Add a SC4List class
This class implements a list type that is binary compatible with the
STLPort list that the Windows version of SC4 was built with.

Unlike the Mac version of the game, the Windows version does not appear
to use a cRZAutoRefCount wrapper for types derived from cIGZUnknown.
Instead it places a pointer to the cIGZUnknown-derived type directly
into the list.
It could have been optimized out by the compiler, or just be something
that Aspyr changed in their port.
Because of this, the SC4List has template code paths for dealing with
cIGZUnknown.