Update SC4HashSet
Changed the type to a class.
Removed the insertion and hash bucket allocation code to fix a memory
leak, SC4 was overwriting our allocated hash buckets with its own.
Updated the destructor to handle freeing the hash buckets that SC4
allocated.
Added a size method and changed the find method to return an iterator.